feat: implementing `internal/client.Pool` to prevent Provider from creating too many connections
This is necessary because Terraform will invoke the Provider's `ConfigureContextFunc` once for each operation a resource requires.
Invoking `NewClient` in there was leading to spawning a connection each time.
The underlying client DOES SUPPORT concurrent use, so we create a _pool_ where each `Client` is identified by a key made of all it's construction parameters.